+++ src/database/bases/HobyDataBase.ts
@@ -1,6 +1,6 @@
-import { Hobby } from "../models/Hobby";
-import BaseDatabase from "./BaseDatabase";
-import { TABLE_HOBBY } from "./tableNames";
+import { Hobby } from "../../models/Hobby";
+import BaseDatabase from "../BaseDatabase";
+import { TABLE_HOBBY } from "../tableNames";
 
 export class HobbyDataBase extends BaseDatabase {
     TABLE_NAME: string = TABLE_HOBBY
