+++ src/database/migrations.ts
@@ -14,14 +14,14 @@
 //         modulo VARCHAR(255) DEFAULT 0
 //     );
     
-//     CREATE TABLE IF NOT EXISTS ${TABLE_ESTUDANTE}(
-//         id VARCHAR(255) PRIMARY KEY,
-//         nome VARCHAR(255) NOT NULL,
-//         email VARCHAR(255) NOT NULL UNIQUE,
-//         data_nasc DATE NOT NULL,
-//         turma_id VARCHAR(255) NOT NULL UNIQUE,
-//         FOREIGN KEY (turma_id) REFERENCES ${TABLE_TURMA}(id)        
-//     );
+    // CREATE TABLE IF NOT EXISTS ${TABLE_ESTUDANTE}(
+    //     id VARCHAR(255) PRIMARY KEY,
+    //     nome VARCHAR(255) NOT NULL,
+    //     email VARCHAR(255) NOT NULL UNIQUE,
+    //     data_nasc DATE NOT NULL,
+    //     turma_id VARCHAR(255) NOT NULL UNIQUE,
+    //     FOREIGN KEY (turma_id) REFERENCES ${TABLE_TURMA}(id)        
+    // );
     
 //     CREATE TABLE IF NOT EXISTS ${TABLE_HOBBY}(
 //         id VARCHAR(255) PRIMARY KEY,
