+++ src/endpoints/hobbys/getHobbys.ts
@@ -1,5 +1,5 @@
 import { Request, Response} from "express";
-import { HobbyDataBase } from "../../database/HobyDataBase";
+import { HobbyDataBase } from "../../database/bases/HobyDataBase";
 
 export const getHobbys = async (req: Request, res: Response) => {
