+++ src/endpoints/turmas/createTurma.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import TurmaDatabase from "../../database/TurmaDatabase";
+import TurmaDatabase from "../../database/bases/TurmaDatabase";
 import { Turma } from "../../models/Turma";
 
 export const createTurma = async (req: Request, res: Response) => {
