+++ src/index.ts
@@ -31,7 +31,7 @@ app.put("/turma/modulo/edit/:id", editModulo)
 
 // Estudante
 
-//
+
 
 //======================================================
 // Hobbys
