+++ src/models/Estudante.ts
@@ -5,11 +5,16 @@ export class Estudante extends Pessoa {
 
     constructor(
         id: string,
-        nome: string,
+        name: string,
         email: string,
         data_nasc: Date,
         turma_id: string,
+        // private hobbies?: string[]
     ) {  
-        super(id, nome, email, data_nasc,turma_id)     
-    }   
+        super(id, name, email, data_nasc,turma_id)     
+    }
+
+    // public getHobbies(){
+    //     return this.hobbies
+    // }    
 }
